+++ components/pages/DetailPage/Cuisines.tsx
