+++ components/pages/DetailPage/DetailPage.tsx
@@ -2,14 +2,14 @@ import React, { useState } from 'react';
 import styled, { css } from 'styled-components';
 import { LazyImage } from 'react-lazy-images';
 
-import Layout from '../../layouts/Layout';
+import Layout from '../../../layouts/Layout';
 
-import Loader from '../core/Loader/Loader';
-import Title from '../core/Title/Title';
-import Button from '../core/Button/Button';
-import BlockTitle from '../core/BlockTitle/BlockTitle';
-import BlockText from '../core/BlockText/BlockText';
-import Rating from '../core/Rating/Rating';
+import Loader from '../../core/Loader/Loader';
+import Title from '../../core/Title/Title';
+import Button from '../../core/Button/Button';
+import BlockTitle from '../../core/BlockTitle/BlockTitle';
+import BlockText from '../../core/BlockText/BlockText';
+import Rating from '../../core/Rating/Rating';
 
 import Timmings from './Timings';
 import Cuisines from './Cuisines';
@@ -17,12 +17,12 @@ import Highlights from './Highlights';
 import Address from './Address';
 import RelatedRestaurants from './RelatedRestaurants';
 
-import { RestaurantDetail, Restaurant } from '../../helpers/types';
-import { getTimmings, getMapSrc } from '../../helpers/utils';
+import { RestaurantDetail, Restaurant } from '../../../helpers/types';
+import { getTimmings, getMapSrc } from '../../../helpers/utils';
 import {
   DETAIL_GENERIC_SRC,
   DEFAULT_TEXT_LOADING,
-} from '../../helpers/staticData';
+} from '../../../helpers/staticData';
 
 type DetailPageProps = {
   data: RestaurantDetail;
