+++ components/pages/DetailPage/RelatedRestaurants.tsx
@@ -1,10 +1,10 @@
 import React from 'react';
 
-import Card from '../core/Card/Card';
+import Card from '../../core/Card/Card';
 
-import { Restaurant } from '../../helpers/types';
+import { Restaurant } from '../../../helpers/types';
 
-import { THUMB_GENERIC_SRC } from '../../helpers/staticData';
+import { THUMB_GENERIC_SRC } from '../../../helpers/staticData';
 
 type RelatedRestaurantsProps = {
   restaurants: Restaurant[];
