+++ components/pages/ErrorPage/ErrorPage.tsx
@@ -2,10 +2,10 @@ import React from 'react';
 import { useRouter } from 'next/router';
 import styled from 'styled-components';
 
-import Layout from '../../layouts/Layout';
+import Layout from '../../../layouts/Layout';
 
-import BlockTitle from '../core/BlockTitle/BlockTitle';
-import Button from '../core/Button/Button';
+import BlockTitle from '../../core/BlockTitle/BlockTitle';
+import Button from '../../core/Button/Button';
 
 const StyledErrorPage = styled.div`
   display: flex;
