+++ components/pages/HomePage/HomePage.tsx
@@ -1,17 +1,17 @@
 import React from 'react';
 import styled from 'styled-components';
 
-import Layout from '../../layouts/Layout';
+import Layout from '../../../layouts/Layout';
 
-import Loader from '../core/Loader/Loader';
-import Title from '../core/Title/Title';
-import Card, { CardType } from '../core/Card/Card';
+import Loader from '../../core/Loader/Loader';
+import Title from '../../core/Title/Title';
+import Card, { CardType } from '../../core/Card/Card';
 
 import {
   HIGHLIGHTED_RESTAURANTS,
   DEFAULT_TEXT_LOADING,
-} from '../../helpers/staticData';
-import { Restaurant } from '../../helpers/types';
+} from '../../../helpers/staticData';
+import { Restaurant } from '../../../helpers/types';
 
 type HomePageProps = {
   isLoading: boolean;
