+++ components/ui/Finder/Finder.tsx
@@ -3,14 +3,14 @@ import { useDispatch } from 'react-redux';
 import { useRouter } from 'next/router';
 import styled from 'styled-components';
 
-import AutocompleteMobile from '../core/Autocomplete/AutocompleteMobile';
-import Autocomplete from '../core/Autocomplete/Autocomplete';
-import Dropdown from '../core/Dropdown/Dropdown';
-import Button from '../core/Button/Button';
+import AutocompleteMobile from '../../core/Autocomplete/AutocompleteMobile';
+import Autocomplete from '../../core/Autocomplete/Autocomplete';
+import Dropdown from '../../core/Dropdown/Dropdown';
+import Button from '../../core/Button/Button';
 
-import useWindowMeasures from '../hooks/useWindowMeasures';
+import useWindowMeasures from '../../hooks/useWindowMeasures';
 
-import { setRelatedRestaurants } from '../../store/actions';
+import { setRelatedRestaurants } from '../../../store/actions';
 
 import {
   DUBLIN_ID,
@@ -19,14 +19,14 @@ import {
   CUISINES,
   THUMB_GENERIC_SRC,
   MIN_RESTAURANTS_LIST,
-} from '../../helpers/staticData';
+} from '../../../helpers/staticData';
 import {
   getFormattedUrlText,
   getCurrentRelatedRestaurants,
-} from '../../helpers/utils';
-import { EntityType, Restaurant } from '../../helpers/types';
+} from '../../../helpers/utils';
+import { EntityType, Restaurant } from '../../../helpers/types';
 
-import { getRestaurants } from '../../services';
+import { getRestaurants } from '../../../services';
 
 type FinderProps = {
   className?: string;
