+++ pages/search/[location]/[cuisine].tsx
@@ -2,8 +2,8 @@ import React, { useRef, useState, useEffect, Dispatch } from 'react';
 import { useSelector, useDispatch } from 'react-redux';
 import { NextPage, NextPageContext } from 'next';
 
-import ErrorPage from '../../../components/ErrorPage/ErrorPage';
-import SearchPage from '../../../components/SearchPage/SearchPage';
+import ErrorPage from '../../../components/pages/ErrorPage/ErrorPage';
+import SearchPage from '../../../components/pages/SearchPage/SearchPage';
 
 import useWindowMeasures from '../../../components/hooks/useWindowMeasures';
 import useScrollPosY from '../../../components/hooks/useScrollPosY';
